Configuring airbnb style rules
