+++ .eslintrc
@@ -1,0 +1,77 @@
+{
+  "extends": "airbnb",
+
+  "rules": {
+    "indent": [2, 4, { "SwitchCase": 1 }],
+    "max-len": [2, 120, 4, {
+      "ignoreUrls": true,
+      "ignoreComments": false
+    }],
+    "jsx-quotes": [2, "prefer-double"],
+    "no-return-assign": 0,
+    "jsx-a11y/no-static-element-interactions": 0,
+    "jsx-a11y/click-events-have-key-events": 0,
+    "class-methods-use-this" :0,
+    "linebreak-style": 0,
+    "react/jsx-boolean-value": 0,
+    "react/require-default-props" :0,
+    "react/no-typos": 0,
+    "react/destructuring-assignment": 0,
+    "react/jsx-one-expression-per-line": 0,
+    "operator-linebreak": 0,
+    "object-curly-newline": 0,
+    "no-throw-literal": 0,
+    "react/jsx-fragments": 0,
+    "react/jsx-indent-props": [2, 4],
+    "react/jsx-indent": [2, 4],
+    "react/jsx-props-no-spreading": 0,
+    "react/jsx-max-props-per-line": [2, {"maximum": 3}],
+    "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
+    "react/jsx-no-bind": [2, {
+      "ignoreRefs": true
+    }],
+    "react/sort-comp": [2, {
+      "order": [
+        "static-methods",
+        "lifecycle",
+        "/^handle.+$/",
+        "/^(get|set)(?!(InitialState$|DefaultProps$|ChildContext$)).+$/",
+        "everything-else",
+        "/^render.+$/",
+        "render"
+      ]
+    }],
+    "prefer-template": "warn",
+    "import/prefer-default-export": "off",
+    "no-console": "off",
+    "arrow-parens": [2, "as-needed"]
+  },
+
+  "env": {
+    "browser": true,
+    "node": true,
+    "es6": true
+  },
+
+  "settings": {
+    "import/resolver": {
+      "node": {
+        "paths": ["src"]
+      }
+    }
+  },
+
+  "globals": {
+    "$": true,
+    "ga": true,
+    "l": true,
+    "nl": true,
+    "sprintf": true
+  },
+
+  "parserOptions": {
+    "ecmaFeatures": {
+      "legacyDecorators": true
+    }
+  }
+}
